+++ scripts/test-installer.ps1
@@ -1,8 +1,8 @@
 param(
     [Parameter(Mandatory = $true)][string]$InstallerPath,
     [Parameter(Mandatory = $true)][string]$PreviousInstallerPath,
-    [string]$ExpectedVersion = '1.3.1',
-    [string]$PreviousVersion = '1.3.0'
+    [string]$ExpectedVersion = '1.3.2',
+    [string]$PreviousVersion = '1.3.1'
 )
 
 $ErrorActionPreference = 'Stop'